feat: implement janitor to clean the cache of expired items
In v1.0.0 expired items are not removed from the cache automatically
contrary to possible expectations. Expired items will remain in the
cache indefinitely until the item is specifically requested to be
removed. The lack of automatically cleaning up expired items from the
cache may lead to confusion and frustration when users are observing the
consumed memory rise while items in the cache are expiring.

To help resolve this I'm introducing the janitor. By default the janitor
is off. Developers may turn it on by configuring a `CleanupInterval` on
the cache, as per the example below. Additionally developers can now
fetch items with expired items by using the `WithExpired` fluent method.

```go
package main

import (
	"fmt"
	"time"

	"github.com/guardian360/go-cache"
)

func main() {
	c := cache.New(
		cache.CleanupInterval(5 * time.Millisecond),
	)
	c.Put("foo", "bar", time.Millisecond)

	// Key "foo" is present in the cache
	fmt.Println(c.WithExpired(true).Items())

	<-time.After(10 * time.Millisecond)

	// Key "foo" is no longer present in the cache
	fmt.Println(c.WithExpired(true).Items())
}
```